move Settings Activity and Fragment to another package
