+++ app/src/androidTest/java/org/mozilla/focus/activity/ChangeLanguageTest.java
@@ -20,6 +20,7 @@ import org.junit.Test;
 import org.mozilla.focus.R;
 import org.mozilla.focus.helper.BeforeTestTask;
 import org.mozilla.focus.utils.AndroidTestUtils;
+import org.mozilla.rocket.settings.SettingsActivity;
 
 import static androidx.test.InstrumentationRegistry.getInstrumentation;
 import static androidx.test.espresso.Espresso.onData;
