+++ app/src/main/java/org/mozilla/rocket/deeplink/task/StartSettingsActivityTask.kt
@@ -1,7 +1,7 @@
 package org.mozilla.rocket.deeplink.task
 
 import android.content.Context
-import org.mozilla.focus.activity.SettingsActivity
+import org.mozilla.rocket.settings.SettingsActivity
 
 class StartSettingsActivityTask(val action: String = "") : Task {
     override fun execute(context: Context) {
